ReferenceElementNode: Move into own file.
